Consolidate FeaturedImageMigrator
